Almost Gemtext scanner
This commit adds a simple scanner for all tokens required by the initial
Almost Gemtext specification.

The scanner uses bufio.Scanner which is easy to implement. However, it
may be possible that the scanner has to inspect some input repeatedly.
This may lead to a slight decrease in performance which I find more than
acceptable for the moment.